requirements.txt
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
Flask-Session==0.8.0
Flask-Migrate==4.1.0
Flask-Limiter==4.1.1
psycopg2-binary==2.9.10
redis==6.2.0
bcrypt==4.1.2
alembic==1.16.4
pytest==8.0.0
hypothesis==6.98.0
pytest-cov==4.1.0
python-dotenv==1.0.0
